backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pandas>=2.2,<3.0
pyarrow>=16.0,<20.0
python-dotenv>=1.0,<2.0
orjson>=3.10,<4.0
